One package for the one binary, and a clippy-clean workspace enforced in CI
The CLI package is now simply `ash` (cargo build -p ash produces the
binary); the library crate it wraps becomes `ash_core`. Every consumer
follows: benchmarks.toml, ash_bench.py, workflows, tests, bench docs.

Clippy: 669 warnings and 5 deny-by-default errors to zero across the
workspace and all targets, now enforced by .github/workflows/lint.yml
with -D warnings. The bulk was mechanical (same-type transmutes and
casts, c-string literals, transmute annotations, dead legacy items,
static_mut_refs rewritten through &raw); bindgen's generated bindings
are lint-allowed at their include sites; the FFI shim crate carries a
documented crate-level allow for missing_safety_doc (its public surface
is the C ABI contract, not a Rust API); and the static-mut modules allow
the two style lints that flag exactly the spelling static_mut_refs
demands. The full test suite and the bench corpus verify behavior is
unchanged (mandelbrot checksum 112798500 intact).

+++ std/src/maps.rs
@@ -347,7 +347,6 @@ pub unsafe extern "C" fn hlp_hballoc() -> *mut hl::hl_hb_map {
 }
 
 #[no_mangle]
-#[allow(clippy::useless_assignment)]
 pub unsafe extern "C" fn hlp_hbset(
     mut m: *mut hl::hl_hb_map,
     key: *mut hl::uchar,
@@ -397,7 +396,7 @@ pub unsafe extern "C" fn hlp_hbset(
         ptr::write(dst, ptr::read(src));
         ptr::write(
             ((*m).cells as *mut i32).wrapping_add(ckey as usize),
-            c as i32,
+            c,
         );
     }
     (*(*m).values.wrapping_add(c as usize)).value = value;
